[Feature] Add GatedDeltaNet all-mode APC (Automatic Prefix Caching) support
Migrate GDN all-mode APC from upstream vllm PR #26807 to vllm-ascend.
This enables SSM recurrent state saving at every block boundary during
prefill for GatedDeltaNet models (Qwen3-Next/Qwen3.5), allowing prefix
cache reuse across requests sharing a common prefix.

Key changes:
- chunk.py/chunk_delta_h.py: Add return_intermediate_states and state_dtype
  params to capture per-chunk SSM states during prefill
- chunk_o.py: Fix dtype cast for state tensor loaded from h buffer
- causal_conv1d.py: Extend causal_conv1d_fn with APC params for
  per-block conv state saving during prefill
- gdn.py: Add _write_apc_ssm_states method and APC branches for both
  interleaved (Qwen3Next) and non-interleaved (Qwen3.5) layouts,
  covering prefill conv1d, prefill SSM, and decode state management
- patch_gdn_attn.py: Inject APC metadata (block indices, chunk metadata)
  into attention metadata when mamba_cache_mode=="all"
- patch_mamba_config.py: Support "all" mode in block size alignment
- patch_qwen3_5.py: Bypass NotImplementedError for all-mode in model init

+++ vllm_ascend/ops/triton/fla/chunk_o.py
@@ -79,7 +79,7 @@ def chunk_fwd_kernel_o(
             # [BK, BT]
             b_k = tl.load(p_k, boundary_check=(0, 1))
             # [BK, BV]
-            b_h = tl.load(p_h, boundary_check=(0, 1))
+            b_h = tl.load(p_h, boundary_check=(0, 1)).to(b_q.dtype)
 
             # [BT, BK] @ [BK, BV] -> [BT, BV]
             b_o += tl.dot(b_q, b_h)
